app/requirements.txt
django
djangorestframework
crypto